+++ packages/opencorvus/script/overlay-snap.ts
@@ -4,7 +4,7 @@
 //   bun run script/overlay-snap.ts <out.png> [url] [w] [h]
 
 import { launchBrowser } from "../../overlay/test/launch"
-import { gotoWithBrowserInactivity } from "./benchmark/browser-inactivity"
+import { gotoWithBrowserInactivity } from "./browser-inactivity"
 import path from "node:path"
 
 const out = process.argv[2]